pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-forecasting-mcp"
version = "0.1.0"
description = "MCP server for financial forecasting using SVM with OpenAI interpretation"
authors = [
    {name = "Financial Forecasting MCP", email = "bhaskar.tripathi@gmail.com"}
]
dependencies = [
    "mcp>=1.0.0",
    "yfinance>=0.2.30",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "ta>=0.10.2",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
financial-forecasting-mcp = "financial_forecasting_mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src" 